Move server code used by main into an importable library called server.
